docker-components/cgaf-datasource/calculate/requirements.txt
psycopg2
pyts
numpy
Pillow
